requirements.txt
psutil
uvicorn
fastapi
paho-mqtt
py-cpuinfo==5.0.0